pyproject.toml
[project]
name = "organice2"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.2",
    "fastapi>=0.134.0",
    "google-api-python-client>=2.190.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "google-cloud-documentai>=3.10.0",
    "google-cloud-resource-manager>=1.16.0",
    "google-cloud-storage>=3.9.0",
    "google-genai>=1.65.0",
    "googlemaps>=4.10.0",
    "httpx>=0.28.1",
    "moviepy>=2.2.1",
    "openai>=2.24.0",
    "opencv-python>=4.13.0.92",
    "pillow>=11.3.0",
    "psutil>=7.2.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.7.4",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "tiktoken>=0.12.0",
    "uvicorn>=0.41.0",
]
